Update the top level interface used to dyntrace promises
This commit updates the experimental setup to use
`dyntrace_promises` function instead of separate
creating, dyntracing and destroying functions.

This interface hides the `dyntracer` object from the
global environment in which the code block is being
traced. This way, the code block is not able to remove
the dyntracer from the environment.