Remove code that creates OK files on successful tracing
If the tracing is successful, a ".OK" file is created
to identify databases which are safe for analysis.
However, due to a recent change in the error reporting
mechanism, this .OK file is created even if the tracing
is unsuccessful due to the vignette throwing an error.
This commit prevents the creation of ".OK" files. This
file is now created by the tracer internally, when the
tracing is successful.