Fix analysis scripts to draw boxplot on violin plots
Some scripts fail during the visualization stage when
`geom_violin` is drawn with `draw_quantiles` argument.
The reason for this is unclear and the error message
is very unhelpful. I believe this happens because these
visualizations tend to be very close to 0 and are bunched
together.
Drawing a boxplot on top of violin plot solves the problem.